refactor(get_redis_clients): move this functionality to a central location
this means we can stop writing the same code in two files while also avoiding circular imports
